requirements.txt
Flask
Flask-PyMongo
pymongo
requests
python-dotenv
pytz